backend/requirements.txt
chromadb
whoosh
minio
networkx
python-dotenv
pyyaml
sentence-transformers
pydantic
fastapi
langchain-community
pypdf
openai
deepagents
python-multipart
uvicorn
langgraph
langchain_openrouter
prometheus-client